src/fly_project/requirements.txt
Django==1.9.1
Markdown==2.6.5
Pillow==3.1.0
PyJWT==1.4.0
defusedxml==0.4.1
django-braces==1.8.1
django-filter==0.12.0
django-oauth-toolkit==0.10.0
djangorestframework==3.3.2
gunicorn==19.4.5
oauthlib==1.0.3
psycopg2==2.6.1
python-social-auth==0.2.14
python3-openid==3.0.9
requests==2.9.1
requests-oauthlib==0.6.0
six==1.10.0
